requirements.txt
absl-py==0.1.10
alabaster==0.7.10
amqp==1.4.9
anaconda-client==1.6.5
anaconda-navigator==1.6.9
anaconda-project==0.8.0
anyjson==0.3.3
argon2-cffi==16.3.0
asn1crypto==0.22.0
astroid==1.5.3
astropy==2.0.2
Babel==2.5.0
backports.shutil-get-terminal-size==1.0.0
backports.weakref==1.0rc1
bcrypt==3.1.3
beautifulsoup4==4.6.0
billiard==3.3.0.23
bitarray==0.8.1
bkcharts==0.2
blaze==0.11.3
bleach==1.5.0
bokeh==0.12.10
boto==2.48.0
Bottleneck==1.2.1
bs4==0.0.1
celery==3.1.26.post2
certifi==2018.1.18
cffi==1.10.0
chardet==3.0.4
click==6.7
cloudpickle==0.4.0
clyent==1.2.2
colorama==0.3.9
colorlover==0.2.1
conda==4.5.0
conda-build==3.0.27
conda-verify==2.0.0
contextlib2==0.5.5
cryptography==2.0.3
cufflinks==0.12.1
cycler==0.10.0
Cython==0.26.1
cytoolz==0.8.2
dask==0.17.2
datashape==0.5.4
decorator==4.1.2
distributed==1.21.6
dj-database-url==0.4.2
Django==1.10.5
django-bootstrap-form==3.3
django-bootstrap3==9.1.0
django-braces==1.12.0
django-celery==3.2.2
django-chartit==0.2.9
django-classy-tags==0.8.0
django-countries==5.0
django-crispy-forms==1.7.0
django-csvimport==2.11
django-custom-user==0.7
django-debug-toolbar==1.9.1
django-extensions==1.9.9
django-filter==1.1.0
django-location-field==2.0.3
django-markup-deprecated==0.0.3
django-misaka==0.2.1
django-model-utils==3.1.1
django-sekizai==0.10.0
django-star-ratings==0.6.0
django-widget-tweaks==1.4.1
djangocms-installer==0.9.7
docutils==0.14
entrypoints==0.2.3
et-xmlfile==1.0.1
Faker==0.7.12
fastcache==1.0.2
filelock==2.0.12
Flask==0.12.2
Flask-Cors==3.0.3
geopy==1.11.0
gevent==1.2.2
glob2==0.5
gmpy2==2.0.8
greenlet==0.4.12
gunicorn==19.7.1
h5py==2.7.0
heapdict==1.0.0
houdini.py==0.1.0
html5lib==0.9999999
idna==2.6
imageio==2.2.0
imagesize==0.7.1
ipykernel==4.6.1
ipython==6.1.0
ipython-genutils==0.2.0
ipywidgets==7.0.0
isort==4.2.15
itsdangerous==0.24
jdcal==1.3
jedi==0.10.2
Jinja2==2.9.6
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.1.0
jupyter-console==5.2.0
jupyter-core==4.3.0
jupyterlab==0.27.0
jupyterlab-launcher==0.4.0
kiwisolver==1.0.1
kombu==3.0.37
lazy-object-proxy==1.3.1
llvmlite==0.20.0
locket==0.2.0
lxml==4.2.1
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==2.2.2
mccabe==0.6.1
misaka==2.1.0
mistune==0.7.4
mkl-fft==1.0.0
mkl-random==1.0.1
mpmath==0.19
msgpack-python==0.4.8
multipledispatch==0.4.9
navigator-updater==0.1.0
nbconvert==5.3.1
nbformat==4.4.0
networkx==2.0
nltk==3.2.5
nose==1.3.7
notebook==5.0.0
numba==0.35.0+10.g143f70e90
numexpr==2.6.2
numpy==1.14.2
numpydoc==0.7.0
odo==0.5.1
olefile==0.44
openpyxl==2.4.8
packaging==16.8
pandas==0.22.0
pandas-datareader==0.6.0
pandocfilters==1.4.2
partd==0.3.8
path.py==10.3.1
pathlib2==2.3.0
patsy==0.4.1
pep8==1.7.0
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.1.1
pkginfo==1.4.1
plotly==2.5.1
ply==3.10
prompt-toolkit==1.0.15
protobuf==3.5.2
psutil==5.4.0
psycopg2==2.7.3.1
ptyprocess==0.5.2
py==1.4.34
pycodestyle==2.3.1
pycosat==0.6.3
pycparser==2.17
pycrypto==2.6.1
pycurl==7.43.0
pyflakes==1.6.0
Pygments==2.2.0
pylint==1.7.4
pyodbc==4.0.17
pyOpenSSL==17.2.0
pyparsing==2.2.0
PySocks==1.6.7
pytest==3.2.1
python-dateutil==2.7.2
pytz==2018.4
PyWavelets==0.5.2
PyYAML==3.12
pyzmq==16.0.2
QtAwesome==0.4.4
qtconsole==4.3.1
QtPy==1.3.1
requests==2.18.4
requests-file==1.4.3
requests-ftp==0.3.1
rope==0.10.5
ruamel-yaml==0.11.14
scikit-image==0.13.0
scikit-learn==0.19.1
scipy==0.19.1
seaborn==0.8.1
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.11.0
snowballstemmer==1.2.1
sortedcollections==0.5.3
sortedcontainers==1.5.7
Sphinx==1.6.3
sphinxcontrib-websupport==1.0.1
spyder==3.2.4
SQLAlchemy==1.2.5
sqlparse==0.2.4
statsmodels==0.8.0
swapper==1.1.0
sympy==1.1.1
tables==3.4.2
tblib==1.3.2
tensorflow==1.5.0
tensorflow-tensorboard==1.5.1
terminado==0.6
testpath==0.3.1
text-unidecode==1.1
toolz==0.8.2
tornado==4.5.2
traitlets==4.3.2
typing==3.6.2
tzlocal==1.4
unicodecsv==0.14.1
urllib3==1.22
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.12.2
whitenoise==3.3.0
widgetsnbextension==3.0.2
wrapt==1.10.11
xlrd==1.1.0
XlsxWriter==1.0.2
xlwt==1.3.0
zict==0.1.3
